fix: moved hapi logic to one place
